subject/contrib/plugins/image_artifact/requirements.txt
python-subjectclient
